Added Score file, moved Score code into Score file, and added functions to charge/penalize players or teams
