+++ lib/player.js
@@ -5,11 +5,12 @@ var Player = {
 		socket.on('set_initials', function(data){
 			if (data && data != socket.initials){
 				Player.update_initials(socket, data);
-				Player.reset_score(socket);
+				Score.reset_score(socket);
 			}
 			io.sockets.emit('user.renamed', socket.initials);
 		});
 	},
+
 	// Broadcast
 	// broadcast who's online boards
 	emit_online_players: function() {
@@ -22,36 +23,6 @@ var Player = {
 			function(results){
 				io.sockets.emit('online', results);
 				Logger.message('Online Player board updated.');
-			});
-	},
-
-	// broadcast high score boards for players
-	emit_high_scores: function() {
-		DB.db_call("SELECT `initials`, TIMESTAMPDIFF(SECOND, connected_at, last_ping) AS `score_seconds`" +
-			" FROM `players`" +
-			" ORDER BY `score_seconds` DESC LIMIT 10", [], "Querying for high scores", function(results){
-				io.sockets.emit('high_scores', results);
-				Logger.message('High Scores board updated.');
-			});
-	},
-
-	// Updates for player
-	// set 0 score (eg for new initials)
-	reset_score: function(socket) {
-		if (!socket || !socket.player_id){
-			Logger.message('Error resetting score: player_id is falsey.');
-			return;
-		}
-
-		DB.db_call("UPDATE `players` SET `connected_at` = CURRENT_TIMESTAMP, `last_ping` = CURRENT_TIMESTAMP WHERE `id` = ?",
-			[socket.player_id],
-			"Resetting score",
-			function(results){
-				if (results.changedRows > 0){
-					Logger.message("player score updated.");
-					socket.start_time = Timer.now();
-					socket.emit('score_minutes', 0);
-				}
 			});
 	},
